Integrate patent-safe HE-AAC v1 (SBR) extension into libfaac
This commit adds a "naive" SBR implementation to libfaac, enabling HE-AAC v1 support while strictly avoiding implementation patents.

Key features:
- API extension with useSbr flag in faacEncConfiguration.
- Dual-rate synchronization: AAC core runs at half-rate (1024 samples) while SBR processes 2048 full-rate samples.
- Optimized 64-band Complex Polyphase Filterbank (QMF) using lookup tables.
- Blind shift spectral transposition with mandatory sign-flipping for aliasing mitigation.
- RMS-based envelope scaling with naive 6-bit quantization.
- Standard-compliant bitstream formatting: SBR data packed in ID_FIL elements (type 13) after core elements.
- Strict CBR compliance for the SBR module via zero-padding.
- SBR disabled for sample rates <= 24kHz and restricted to Mono/Stereo.

Fixes in v2:
- Fixed M_PI undeclared identifier on Windows.
- Fixed numerous compiler warnings (signed/unsigned mismatch, loss of data) to satisfy strict CI environments.
- Corrected time-alignment and dual-rate buffering.
- Improved SBR bitstream compliance with mandatory fields.

The implementation is syntactically compliant with ISO/IEC 14496-3 and decodable by reference HE-AAC decoders.

+++ libfaac/bitstream.c
@@ -866,7 +866,7 @@ static int WriteSBRData(faacEncStruct* hEncoder,
 {
     sbr_info_t *sbr = (sbr_info_t *)hEncoder->sbr_info;
     int bits = 0;
-    int sbr_frame_bits = (SBR_BPC * hEncoder->numChannels * FRAME_LEN) / hEncoder->inputSampleRate;
+    long sbr_frame_bits = (long)(SBR_BPC * hEncoder->numChannels * FRAME_LEN) / (long)hEncoder->inputSampleRate;
     long start_bits = bitStream->numBit;
 
     if (writeFlag) {
@@ -874,7 +874,7 @@ static int WriteSBRData(faacEncStruct* hEncoder,
         /* count is 4 bits, if count == 15, then 8 more bits */
         /* For simplicity, we assume the total SBR payload fits in the standard fill element. */
         /* HE-AAC SBR normally uses extension_type 13. */
-        int count = (sbr_frame_bits / 8);
+        int count = (int)(sbr_frame_bits / 8);
         if (count < 15) {
             PutBit(bitStream, count, 4);
         } else {
